requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
Faker==4.14.2
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
validate-docbr==1.8.2
django-cors-headers==3.5.0
validators==0.20.0
decorator==5.1.1
mysqlclient==2.1.1
django-filter==2.3.0